Feat(lms) : Step2 : 수강신청(도메인 모델)
- BaseEntity 분리
- Course 도메인 수정
